requirements.txt
streamlit
groq
python-dotenv
pydantic
pandas
requests
asyncio
pathlib